Pull workspace discovery up
